+++ arch/arm/mm/pgd.c
@@ -24,7 +24,7 @@
 #define __pgd_free(pgd)	kfree(pgd)
 #else
 #define __pgd_alloc()	(pgd_t *)__get_free_pages(GFP_KERNEL | __GFP_REPEAT, 2)
-#define __pgd_free(pgd)	free_pages((unsigned long)pgd, 2)
+#define __pgd_free(pgd)	free_pages(pgd, 2)
 #endif
 
 /*
