+++ arch/cris/arch-v32/drivers/pci/dma.c
@@ -45,6 +45,6 @@ void dma_free_coherent(struct device *dev, size_t size,
 	int order = get_order(size);
 
 	if (!dma_release_from_coherent(dev, order, vaddr))
-		free_pages((unsigned long)vaddr, order);
+		free_pages(vaddr, order);
 }
